move processExperimentalAssetsArg to experimental-assets.ts
