+++ packages/wrangler/src/dev.tsx
@@ -1,6 +1,5 @@
 import assert from "node:assert";
 import events from "node:events";
-import { existsSync } from "node:fs";
 import path from "node:path";
 import util from "node:util";
 import { isWebContainer } from "@webcontainer/env";
@@ -22,7 +21,7 @@ import registerDevHotKeys from "./dev/hotkeys";
 import { maybeRegisterLocalWorker } from "./dev/local";
 import { startDevServer } from "./dev/start-server";
 import { UserError } from "./errors";
-import { getExperimentalAssetsBasePath } from "./experimental-assets";
+import { processExperimentalAssetsArg } from "./experimental-assets";
 import { run } from "./experimental-flags";
 import isInteractive from "./is-interactive";
 import { logger } from "./logger";
@@ -804,31 +803,8 @@ export async function startDev(args: StartDevOptions) {
 			});
 		}
 
-		const experimentalAssets = args.experimentalAssets
-			? { directory: args.experimentalAssets }
-			: config.experimental_assets;
+		const experimentalAssets = processExperimentalAssetsArg(args, config);
 		if (experimentalAssets) {
-			const experimentalAssetsBasePath = getExperimentalAssetsBasePath(
-				config,
-				args.experimentalAssets
-			);
-			const resolvedExperimentalAssetsPath = path.resolve(
-				experimentalAssetsBasePath,
-				experimentalAssets.directory
-			);
-
-			if (!existsSync(resolvedExperimentalAssetsPath)) {
-				const sourceOfTruthMessage = args.experimentalAssets
-					? '"--experimental-assets" command line argument'
-					: '"experimental_assets.directory" field in your configuration file';
-
-				throw new UserError(
-					`The directory specified by the ${sourceOfTruthMessage} does not exist:\n` +
-						`${resolvedExperimentalAssetsPath}`
-				);
-			}
-
-			experimentalAssets.directory = resolvedExperimentalAssetsPath;
 			args.forceLocal = true;
 		}
